Remove remaining processors + adapt tests
