--- x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/gen/processor/ConstantNamedWriteable.java
@@ -1,17 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-package org.elasticsearch.xpack.esql.core.expression.gen.processor;
-
-import org.elasticsearch.common.io.stream.NamedWriteable;
-
-/**
- * Marker interface used by QL for pluggable constant serialization.
- */
-public interface ConstantNamedWriteable extends NamedWriteable {
-
-}
