+++ x-pack/plugin/esql-core/src/main/java/org/elasticsearch/xpack/esql/core/expression/predicate/operator/comparison/In.java
@@ -89,7 +89,20 @@ public class In extends ScalarFunction {
         if (Expressions.isNull(value) || list.size() == 1 && Expressions.isNull(list.get(0))) {
             return null;
         }
-        return InProcessor.apply(value.fold(), foldAndConvertListOfValues(list, value.dataType()));
+        return apply(value.fold(), foldAndConvertListOfValues(list, value.dataType()));
+    }
+
+    private static Boolean apply(Object input, List<Object> values) {
+        Boolean result = Boolean.FALSE;
+        for (Object v : values) {
+            Boolean compResult = Comparisons.eq(input, v);
+            if (compResult == null) {
+                result = null;
+            } else if (compResult == Boolean.TRUE) {
+                return Boolean.TRUE;
+            }
+        }
+        return result;
     }
 
     @Override
